+++ core/domain/interface/merchant/merchant_repo.go
@@ -16,10 +16,10 @@ type IMerchantRepo interface {
 	CreateMerchant(*Merchant) IMerchant
 
 	// 创建会员申请商户密钥
-	CreateSignUpToken(memberId int32) string
+	CreateSignUpToken(memberId int64) string
 
 	// 根据商户申请密钥获取会员编号
-	GetMemberFromSignUpToken(token string) int32
+	GetMemberFromSignUpToken(token string) int64
 
 	// 获取商户的编号
 	GetMerchantsId() []int32
