+++ core/domain/member/gift_card.go
@@ -18,11 +18,11 @@ import (
 var _ member.IGiftCardManager = new(giftCardManagerImpl)
 
 type giftCardManagerImpl struct {
-	memberId int32
+	memberId int64
 	rep      member.IMemberRepo
 }
 
-func newGiftCardManagerImpl(memberId int32, rep member.IMemberRepo) member.IGiftCardManager {
+func newGiftCardManagerImpl(memberId int64, rep member.IMemberRepo) member.IGiftCardManager {
 	return &giftCardManagerImpl{
 		memberId: memberId,
 		rep:      rep,
